Move spec file to __test__
