+++ src/__tests__/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import App from "./App";
+import App from "../App";
 
 Enzyme.configure({ adapter: new Adapter() });
